+++ sdk/src/main/java/ly/count/android/sdk/Event.java
@@ -50,6 +50,8 @@ class Event {
     protected static final String PV_ID_KEY = "pvid";
     protected static final String CV_ID_KEY = "cvid";
     protected static final String PE_ID_KEY = "peid";
+    protected static final String CV_NAME_KEY = "cly_cvn";
+    protected static final String PV_NAME_KEY = "cly_pvn";
 
     public String key;
     public Map<String, Object> segmentation;
@@ -63,6 +65,8 @@ class Event {
     public String pvid;
     public String cvid;
     public String peid;
+    public String cly_cvn;
+    public String cly_pvn;
 
     Event() {
     }
@@ -104,6 +108,14 @@ class Event {
 
             if (peid != null) {
                 json.put(PE_ID_KEY, peid);
+            }
+
+            if (cly_cvn != null) {
+                json.put(CV_NAME_KEY, cly_cvn);
+            }
+
+            if (cly_pvn != null) {
+                json.put(PV_NAME_KEY, cly_pvn);
             }
 
             JSONObject jobj = new JSONObject();
@@ -176,6 +188,14 @@ class Event {
                 event.peid = json.getString(PE_ID_KEY);
             }
 
+            if (!json.isNull(CV_NAME_KEY)) {
+                event.cly_cvn = json.getString(CV_NAME_KEY);
+            }
+
+            if (!json.isNull(PV_NAME_KEY)) {
+                event.cly_pvn = json.getString(PV_NAME_KEY);
+            }
+
             if (!json.isNull(SEGMENTATION_KEY)) {
                 //we would also enter here if segmentation was set to an empty object
                 JSONObject segm = json.getJSONObject(SEGMENTATION_KEY);
@@ -230,6 +250,8 @@ class Event {
             Objects.equals(pvid, e.pvid) &&
             Objects.equals(cvid, e.cvid) &&
             Objects.equals(peid, e.peid) &&
+            Objects.equals(cly_cvn, e.cly_cvn) &&
+            Objects.equals(cly_pvn, e.cly_pvn) &&
             Objects.equals(segmentation, e.segmentation);
     }
 
@@ -241,6 +263,8 @@ class Event {
             (pvid != null ? pvid.hashCode() : 1) ^
             (cvid != null ? cvid.hashCode() : 1) ^
             (peid != null ? peid.hashCode() : 1) ^
+            (cly_cvn != null ? cly_cvn.hashCode() : 1) ^
+            (cly_pvn != null ? cly_pvn.hashCode() : 1) ^
             (timestamp != 0 ? (int) timestamp : 1);
     }
 }