+++ sdk/src/main/java/ly/count/android/sdk/ModuleEvents.java
@@ -17,12 +17,11 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
 
     //used for tracking recorded custom event ID's. This is not updated when internal events are recorded
     String previousEventId = "";
-
     EventQueueProvider eventQueueProvider;
     ViewIdProvider viewIdProvider;
-
     SafeIDGenerator safeEventIDGenerator;
-    boolean visibilityTracking;
+    private final boolean viewNameRecordingEnabled;
+    private final boolean visibilityTracking;
 
     ModuleEvents(Countly cly, CountlyConfig config) {
         super(cly, config);
@@ -32,6 +31,7 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
         config.eventProvider = this;
         eventQueueProvider = config.eventQueueProvider;
         safeEventIDGenerator = config.safeEventIDGenerator;
+        viewNameRecordingEnabled = config.experimental.viewNameRecordingEnabled;
         visibilityTracking = config.experimental.visibilityTrackingEnabled;
 
         eventsInterface = new Events();
@@ -118,10 +118,20 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
         String pvid = null; // Previous View ID
         String cvid = null; // Current View ID
 
+        String previousViewName = null;
+        String currentViewName = null;
+
+        if (viewNameRecordingEnabled) {
+            previousViewName = _cly.moduleViews.previousViewName;
+        }
+
         if (key.equals(ModuleViews.VIEW_EVENT_KEY)) {
             pvid = viewIdProvider.getPreviousViewId();
         } else {
             cvid = viewIdProvider.getCurrentViewId();
+            if (viewNameRecordingEnabled) {
+                currentViewName = _cly.moduleViews.currentViewName;
+            }
         }
 
         if (pcc != null) {
@@ -135,13 +145,13 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
             case ModuleFeedback.NPS_EVENT_KEY:
             case ModuleFeedback.SURVEY_EVENT_KEY:
                 if (consentProvider.getConsent(Countly.CountlyFeatureNames.feedback)) {
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(true);
                 }
                 break;
             case ModuleFeedback.RATING_EVENT_KEY: //these events can be reported from a lot of sources, therefore multiple consents could apply
                 if (consentProvider.getConsent(Countly.CountlyFeatureNames.starRating) || consentProvider.getConsent(Countly.CountlyFeatureNames.feedback)) {
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(false);
                 }
                 break;
@@ -151,19 +161,19 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
                         segmentation = new HashMap<>();
                     }
                     addVisibilityToSegmentation(segmentation);
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(false);
                 }
                 break;
             case ModuleViews.ORIENTATION_EVENT_KEY:
                 if (consentProvider.getConsent(Countly.CountlyFeatureNames.users)) {
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(false);
                 }
                 break;
             case ModulePush.PUSH_EVENT_ACTION:
                 if (consentProvider.getConsent(Countly.CountlyFeatureNames.push)) {
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(true);
                 }
                 break;
@@ -172,7 +182,7 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
                     if (segmentation != null) {
                         UtilsInternalLimits.removeUnsupportedDataTypes(segmentation, L);
                     }
-                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null);
+                    eventQueueProvider.recordEventToEventQueue(key, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, null, previousViewName, currentViewName);
                     _cly.moduleRequestQueue.sendEventsIfNeeded(false);
                 }
                 break;
@@ -184,7 +194,7 @@ public class ModuleEvents extends ModuleBase implements EventProvider {
                     }
                     UtilsInternalLimits.applySdkInternalLimitsToSegmentation(segmentation, _cly.config_.sdkInternalLimits, L, "[ModuleEvents] recordEventInternal");
                     addVisibilityToSegmentation(segmentation);
-                    eventQueueProvider.recordEventToEventQueue(keyTruncated, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, previousEventId);
+                    eventQueueProvider.recordEventToEventQueue(keyTruncated, segmentation, count, sum, dur, timestamp, hour, dow, eventId, pvid, cvid, previousEventId, previousViewName, currentViewName);
                     previousEventId = eventId;
                     _cly.moduleRequestQueue.sendEventsIfNeeded(false);
                 }